requirements.txt
rembg==2.0.50
Pillow==9.3.0
streamlit==1.21.0
numba==0.57
numpy==1.23.0
altair
pandas
google-api-python-client
google-auth
google-auth-httplib2
protobuf
click==8
opencv-python-headless
blind-watermark
fuzzywuzzy
pywavelets

